# faithgpt-platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: FaithGPT API
  version: "1.0.0"
  description: >
    Canonical REST API contract for FaithGPT™ ("Transforming Scripture into Daily Living").
    Conventions, envelopes, pagination, streaming, and rate-limiting are documented narratively
    in /faithgpt-platform/docs/09-api-architecture.md — this document is the reviewable
    machine-readable snapshot. Schema names mirror /faithgpt-platform/schema/schema.prisma
    verbatim (see /faithgpt-platform/docs/00-shared-reference.md §13).
  contact:
    name: FaithGPT Engineering
    url: https://faithgpt.app
servers:
  - url: https://api.faithgpt.app/api/v1
    description: Production
  - url: https://staging-api.faithgpt.app/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =====================================================================================
# PATHS
# =====================================================================================
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get the authenticated user's profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeUser'
    patch:
      tags: [Users]
      summary: Update the authenticated user's profile/preferences
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                avatarUrl: { type: string, nullable: true }
                denominationLens: { type: string, nullable: true }
                locale: { type: string }
                timezone: { type: string }
                defaultBibleVersionId: { type: string, nullable: true }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeUser'
    delete:
      tags: [Users]
      summary: Soft-delete the authenticated user's account (14-day recoverable, §12 §5)
      responses:
        "202":
          description: Account scheduled for deletion
...
  /users/me/sessions/{id}:
    delete:
      tags: [Users]
      summary: Revoke a single session ("log out this device")
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "204":
          description: Session revoked
...
  /images:
    get:
      tags: [Visual Studio]
      summary: List "My Images"
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Image generations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageGeneration'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
  /images/generate:
    post:
      tags: [Visual Studio]
      summary: Queue an AI Scripture Image Generator™ job (async, BullMQ `image-generation` queue)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateImageRequest'
      responses:
        "202":
          description: Job queued — poll GET /images/{id} or await push notification
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ImageGeneration'
        "429":
          $ref: '#/components/responses/QuotaExceeded'
        "451":
          description: Blocked by pre-generation content filter (§19 §1)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /images/{id}:
    get:
      tags: [Visual Studio]
      summary: Get image generation status/result
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Image generation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ImageGeneration'
        "404":
          $ref: '#/components/responses/NotFound'
  /story-visualizations:
    post:
      tags: [Visual Studio]
      summary: Generate an AI Bible Story Visualizer™ sequence (async, multi-frame)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storyKey, style, outputType]
              properties:
                storyKey: { type: string, example: "david-and-goliath" }
                style:
                  $ref: '#/components/schemas/ImageStyle'
                outputType:
                  type: string
                  enum: [STORYBOARD, COMIC, STORYBOOK, TEACHING_SLIDES]
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BibleStoryVisualization'
  /story-visualizations/{id}:
    get:
      tags: [Visual Studio]
      summary: Get a story visualization with its frames
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Story visualization with frames
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BibleStoryVisualizationDetail'
  /memory-verse-cards:
    post:
      tags: [Visual Studio]
      summary: Generate a Memory Verse Visualizer™ card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verseId, format]
              properties:
                verseId: { type: string }
                format:
                  $ref: '#/components/schemas/MemoryCardFormat'
                style:
                  $ref: '#/components/schemas/ImageStyle'
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MemoryVerseCard'
    get:
      tags: [Visual Studio]
      summary: List "My Memory Verse Cards"
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Memory verse cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemoryVerseCard'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'

  # ---------------------------------------------------------------------------------
  # JOURNAL
  # ---------------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Opaque cuid()
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    QuotaExceeded:
      description: Tier-based AI usage quota exceeded (429 RATE_LIMITED, code=QUOTA_EXCEEDED in details)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
            details:
              type: object
              nullable: true
              additionalProperties: true
    CursorPaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
...
    EnvelopeUser:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/User'
...
    UserRole:
      type: string
      enum: [USER, GROUP_LEADER, MINISTRY_ADMIN, MODERATOR, CONTENT_EDITOR, SUPPORT, ADMIN, SUPER_ADMIN]
    SubscriptionTier:
      type: string
      enum: [FREE, PLUS, PREMIUM, MINISTRY, LIFETIME]
...
    ImageStyle:
      type: string
      enum: [PHOTOREALISTIC, CINEMATIC, CHRISTIAN_ARTWORK, CARTOON, CHILDRENS_STORYBOOK, COMIC_BOOK, WATERCOLOR, OIL_PAINTING, MINIMALIST_POSTER, SOCIAL_MEDIA_GRAPHIC]
    ImageSourceType:
      type: string
      enum: [VERSE, CHAPTER, BIBLE_STORY, DEVOTION, PRAYER, SERMON, MEMORY_VERSE, CUSTOM_PROMPT]
    GenerationStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, MODERATION_BLOCKED]
    MemoryCardFormat:
      type: string
      enum: [FLASHCARD, WALLPAPER, LOCK_SCREEN, CHILDRENS_CARD, PRINTABLE]
...
    ModerationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, FLAGGED]
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        role:
          $ref: '#/components/schemas/UserRole'
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        denominationLens: { type: string, nullable: true, example: "NEUTRAL" }
        locale: { type: string, example: "en" }
        timezone: { type: string, example: "UTC" }
        defaultBibleVersionId: { type: string, nullable: true }
        emailVerifiedAt: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    GenerateImageRequest:
      type: object
      required: [sourceType, style]
      properties:
        sourceType:
          $ref: '#/components/schemas/ImageSourceType'
        sourceKey: { type: string, nullable: true, description: "passageKey, storyKey, etc. (required unless sourceType=CUSTOM_PROMPT)" }
        devotionId: { type: string, nullable: true }
        prayerId: { type: string, nullable: true }
        sermonId: { type: string, nullable: true }
        style:
          $ref: '#/components/schemas/ImageStyle'
        userPrompt: { type: string, nullable: true }
    ImageGeneration:
      type: object
      properties:
        id: { type: string }
        sourceType:
          $ref: '#/components/schemas/ImageSourceType'
        sourceKey: { type: string }
        style:
          $ref: '#/components/schemas/ImageStyle'
        refinedPrompt: { type: string }
        status:
          $ref: '#/components/schemas/GenerationStatus'
        resultUrl: { type: string, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        isWatermarked: { type: boolean }
        moderationStatus:
          $ref: '#/components/schemas/ModerationStatus'
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    BibleStoryVisualization:
      type: object
      properties:
        id: { type: string }
        storyKey: { type: string, example: "david-and-goliath" }
        style:
          $ref: '#/components/schemas/ImageStyle'
        outputType: { type: string, enum: [STORYBOARD, COMIC, STORYBOOK, TEACHING_SLIDES] }
        createdAt: { type: string, format: date-time }
    BibleStoryVisualizationDetail:
      allOf:
        - $ref: '#/components/schemas/BibleStoryVisualization'
        - type: object
          properties:
            frames:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  sequence: { type: integer }
                  caption: { type: string }
                  imageGeneration:
                    $ref: '#/components/schemas/ImageGeneration'
    MemoryVerseCard:
      type: object
      properties:
        id: { type: string }
        verseId: { type: string }
        format:
          $ref: '#/components/schemas/MemoryCardFormat'
        imageGeneration:
          $ref: '#/components/schemas/ImageGeneration'
        createdAt: { type: string, format: date-time }

    # --- Journal ---------------------------------------------------------
...